Move expand from coeff to whir
