+++ witness/src/lib.rs
@@ -1,7 +1,6 @@
 use multilinear_extensions::mle::{DenseMultilinearExtension, IntoMLE};
 use p3::{
-    dft::{Radix2DitParallel, TwoAdicSubgroupDft},
-    field::{Field, PrimeCharacteristicRing, TwoAdicField},
+    field::{Field, PrimeCharacteristicRing},
     matrix::Matrix,
 };
 use rand::{Rng, distributions::Standard, prelude::Distribution};
@@ -70,6 +69,12 @@ impl<T: Sized + Sync + Clone + Send + Copy + Default + PrimeCharacteristicRing> 
         let mut inner = self.inner;
         inner.pad_to_height(padded_height, T::default());
         inner
+    }
+
+    pub fn set_num_rows_to_height(&mut self, new_height: usize, fill: T) {
+        assert!(new_height >= self.height());
+        self.inner.pad_to_height(new_height, fill);
+        self.num_rows = new_height;
     }
 
     pub fn n_col(&self) -> usize {
@@ -253,19 +258,4 @@ impl<F: Sync + Send + Copy + PrimeCharacteristicRing> Index<usize> for RowMajorM
         let num_col = self.n_col();
         &self.inner.values[num_col * idx..][..num_col]
     }
-}
-
-pub fn expand_from_coeff<F: Ord + TwoAdicField>(
-    mut coeffs: RowMajorMatrix<F>,
-    expansion: usize,
-) -> RowMajorMatrix<F> {
-    let expanded_size = coeffs.height() * expansion;
-    coeffs.pad_to_height(expanded_size, F::ZERO);
-    coeffs.num_rows = expanded_size;
-    let dft = Radix2DitParallel::<F>::default();
-    let m = coeffs.into_default_padded_p3_rmm().to_row_major_matrix();
-    RowMajorMatrix::new_by_inner_matrix(
-        dft.dft_batch(m).to_row_major_matrix(),
-        InstancePaddingStrategy::Default,
-    )
 }
